requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==2.3.7
openpyxl==3.1.5
psycopg2-binary==2.9.9
gunicorn==22.0.0
